+++ include/Gyroscope/Gyroscope_GY_521_mock.h
@@ -12,18 +12,18 @@ class Gyroscope_GY_521_mock:public Gyroscope{
 
     int _teardown_gyroscope_sensor();
 
-    double _get_x_rotation();
+    float _get_x_rotation();
 
-    double _get_y_rotation();
+    float _get_y_rotation();
 
-    double _get_z_rotation();
+    float _get_z_rotation();
 
     /**
      * Allow to load the values directly from a file
      * It can manage values refreshment when needed
      * this function can't be common between gyroscope and accelerometer values because of static values
      */
-    int load_from_file(double *buffer, axis axis_desired);
+    int load_from_file(float *buffer, axis axis_desired);
 
 public:
     using Gyroscope::Gyroscope;
