+++ source/Accelerometer/Accelerometer_GY_521_mock.cpp
@@ -37,20 +37,20 @@ int Accelerometer_GY_521_mock::_teardown_accelerometer_sensor(){
     return error_code;
 }
 
-double Accelerometer_GY_521_mock::_get_x_acceleration(){
-    double value = 0.0;
+float Accelerometer_GY_521_mock::_get_x_acceleration(){
+    float value = 0.0;
     load_from_file(&value, AXIS_X);
     return value;
 }
 
-double Accelerometer_GY_521_mock::_get_y_acceleration(){
-    double value  = 0.0;
+float Accelerometer_GY_521_mock::_get_y_acceleration(){
+    float value  = 0.0;
     load_from_file(&value, AXIS_Y);
     return value;
 }
 
-double Accelerometer_GY_521_mock::_get_z_acceleration(){
-    double value = 0.0;
+float Accelerometer_GY_521_mock::_get_z_acceleration(){
+    float value = 0.0;
     load_from_file(&value, AXIS_Z);
     return value;
 }
@@ -60,14 +60,14 @@ double Accelerometer_GY_521_mock::_get_z_acceleration(){
  * It can manage values refreshment when needed
  * this function can't be common between gyroscope and accelerometer values because of static values
  */
-int Accelerometer_GY_521_mock::load_from_file(double *buffer, axis axis_desired){
+int Accelerometer_GY_521_mock::load_from_file(float *buffer, axis axis_desired){
     static bool x_new = true;
     static bool y_new = true;
     static bool z_new = true;
-    static double x_value = 0.0;
-    static double y_value = 0.0;
-    static double z_value = 0.0;
-    double values_array[3];
+    static float x_value = 0.0;
+    static float y_value = 0.0;
+    static float z_value = 0.0;
+    float values_array[3];
     char *values = NULL;
     size_t len;
     // If the value has already been fetched, we must triger a new line reading
